requirements.txt
selenium
webdriver-manager
google-generativeai
colorama
pandas
matplotlib
fake-useragent
python-dotenv